🏗️ Save type declarations in separate files
